pyproject.toml
[project]
name = "fastflow_api"
version = "0.1.0"
description = "Framework Backend custom"
authors = [
    { name = "Abdul Rohman Masrifan", email = "masrifan26@gmail.com" }
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.26.0",
    "pendulum>=3.0.0",
    "sqlalchemy>=2.0.25",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3.post1",
    "pydantic[email]>=2.5.3",
    "pymysql>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pyotp>=2.9.0",
    "cityhash>=0.4.7",
    "pytest-bdd>=7.0.1",
    "passlib>=1.7.4",
    "fastapi-mail>=1.4.1",
    "ua-parser>=0.18.0",
    "requests>=2.31.0",
    "pillow>=10.2.0",
    "aiofiles>=23.2.1",
    "aiobotocore[boto3]>=2.9.1",
    "qrcode[pil]>=7.4.2",
    "apscheduler>=3.10.1",
    "httpx>=0.27.2",
    "gunicorn>=21.0.0",
    "python-socketio (>=5.13.0,<6.0.0)",
    "python-barcode (>=0.15.1,<0.16.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "pyttsx3 (>=2.98,<3.0)",
    "pydub (>=0.25.1,<0.26.0)"
]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
